+++ change-log.md
@@ -2,6 +2,51 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 
+## [6.2.0]
+
+### Added
+
+- Form Write now supports fields as well as properties.
+
+## [6.1.0]
+
+### Added
+
+- `Konsole.Platform.Windows.HighSpeedWriter` (native windows driver) now included and no longer an external package.
+
+## [6.0.0]
+
+### Added
+
+- new `IConsole` interfaces for simpler abstractions, see new contract docs [here]
+
+### fixed
+
+- #40 WriteLine conflicts with String.Format, support string containing {0} {json} tokens.
+- #41 If a `Write` ends exactly on the last column, then the cursor is not advanced to the next line. 
+
+## [5.4.4]
+
+### Added
+
+- Printing objects with nullable fields now prints "Null" for any nullable fields.
+
+### Fixed
+
+- #39 Add double to NumericTypes in FieldReader.
+
+## [5.4.3]
+
+### Fixed
+
+- #43 Form.Write(null) throws exception, new Form().Write(null); should not throw exception. Instead it should write "Null"
+
+## [5.4.2]
+
+### Fixed
+
+- `WhenOpeningInlineShould_open_window_at_current_cursorTop()` : OpenBox (inline) when supplying only a width and a height was not opening the new inline window at the current cursor position.
+
 ## [5.4.1]
 
 ### Fixed